Added resources and resource manager
Resources can now be loaded through the ResourceManager class. This
ensures that the resources are loaded from files only once, the next
time they are accessed, a copy will be returned. The copy returned is
refcounted, so that it is automatically removed when it's no longer
needed.
Resources created from code and not from files are also given a name so
that they can be easily accesssed from anywhere in the code